python-validation-loop/requirements.txt
# Python Dependencies for Audio Validation Loop

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# LLM (Gemini)
google-generativeai>=0.3.0

# Fish Audio TTS (install from their repo)
# fish-audio-sdk  # Install from: https://github.com/fishaudio/fish-speech

# Optional: Syllable counting
pyphen>=0.14.0

# Utilities
numpy>=1.24.0
